Build fat jar along with jlinking
What
===
Build fat jar along with jlinking.

Why
===
In 9a02922e98d84bb1ea0b19ff8d2259b5cb7b0e9c the fat jar was removed and
replaced with scripts to jlink a self-contained distribution. While this
is very cool new functionality, a self contained fat jar can be more
convenient and it would be helpful for users where a fat jar is simpler
if one was still generated. This is especially helpful for users of the
java-language-server who are using it standalone.