Make javaLsFlag.txt and lib/src.zip optional
What
===
Make javaLsFlag.txt and lib/src.zip optional.

Why
===
When using this tool standalone the lib/src.zip file doesn't seem to be
packaged inside what is jlinked and exceptions are raised because it
cannot be found. From the tests I've run it doesn't seem to make any
difference if the zip file is present or not, at least to the features I
am using. I don't really understand what the txt and zip file are for,
but it seems like this language server
protocol can be used without them.