Fail quietly on MalformedInputException when parsing package name
When searching for Java files, it's possible to find a `.java` file with
invalid contents, such as in the case of a Vim undo file. When this
happens, a MalformedInputException is thrown, causing initialization to
fail. In this case, it would be better to quietly ignore the failure
than to abort entirely.