Split Config handlers
This allows for better dealing with maven, gradel or bazel as sources for config.

Also, this includes a configuration for bazel to allow for switching binary (i.e. for selecting bazel version or for using bazelisk as frontend for bazel).

Maven version was extracted but not hooked, further work needs to be done there.

Gradle version was not extracted;

`InferConfig` was kept, but adapted to use the same interface;

current behavior as of this commit is that `BazelConfig` will be used IFF `{"bazel": "..."}` is set in the configs.